+++ plugins/module_utils/api_endpoints.py
@@ -2,12 +2,13 @@ from enum import Enum
 
 
 class CDOAPI(Enum):
-    DEVICES = "aegis/rest/v1/services/targets/devices"
-    WORKSET = "aegis/rest/v1/services/common/workingset"
+    SVCS = "aegis/rest/v1/services"
+    DEVICES = f"{SVCS}/targets/devices"
+    WORKSET = f"{SVCS}/common/workingset"
+    FTDS = f"{SVCS}/firepower/ftds"
+    FMC = f"{SVCS}/fmc/appliance"
+    OBJS = f"{SVCS}/targets/objects"
+    LARS = f"{SVCS}/targets/proxies"
+    ASA_CONFIG = f"{SVCS}/asa/configs"
     SPECIFIC_DEVICE = "aegis/rest/v1/device/{uid}/specific-device"
-    LARS = "aegis/rest/v1/services/targets/proxies"
-    ASA_CONFIG = "aegis/rest/v1/services/asa/configs"
-    IOS_CONFIG = ""
     FMC_ACCESS_POLICY = "fmc/api/fmc_config/v1/domain/{domain_uid}/policy/accesspolicies"
-    FTDS = "aegis/rest/v1/services/firepower/ftds"
-    FMC = "aegis/rest/v1/services/fmc/appliance"