+++ plugins/module_utils/requests.py
@@ -1,22 +1,37 @@
 # -*- coding: utf-8 -*-
 from __future__ import absolute_import, division, print_function
+
 __metaclass__ = type
 
 import requests
+import urllib.parse
 from enum import Enum
 from functools import wraps
 from .errors import DuplicateObject, APIError, DeviceNotFound
 
+# fmt: off
+# Remove for publishing....
+import logging
+logging.basicConfig()
+logger = logging.getLogger('requests')
+fh = logging.FileHandler('/tmp/requests.log')
+fh.setLevel(logging.DEBUG)
+logger.setLevel(logging.DEBUG)
+logger.addHandler(fh)
+logger.debug("Query Logger started......")
+# fmt: on
+
 
 class CDORegions(Enum):
-    """ CDO API Endpoints by Region"""
-    us = 'www.defenseorchestrator.com'
-    eu = 'www.defenseorchestrator.eu'
-    apj = 'apj.cdo.cisco.com'
+    """CDO API Endpoints by Region"""
+
+    us = "www.defenseorchestrator.com"
+    eu = "www.defenseorchestrator.eu"
+    apj = "apj.cdo.cisco.com"
 
     @classmethod
     def get_endpoint(region: object, input_region: str) -> str:
-        """ Given a region (input_region) , return the endpoint"""
+        """Given a region (input_region) , return the endpoint"""
         return region[input_region].value
 
 
@@ -52,16 +67,26 @@ class CDORequests:
     def create_session(token: str, version: str) -> str:
         """Helper function to set the auth token and accept headers in the API request"""
         http_session = requests.Session()
-        http_session.headers = {"Authorization": f"Bearer {token.strip()}", "Accept": "*/*",
-                                "Content-Type": "application/json", "User-Agent": f"AnsibleCDOModule/{version}"}
+        http_session.headers = {
+            "Authorization": f"Bearer {token.strip()}",
+            "Accept": "*/*",
+            "Content-Type": "application/json",
+            "User-Agent": f"AnsibleCDOModule/{version}",
+        }
         return http_session
 
     @CDOAPIWrapper()
     @staticmethod
     def get(http_session: requests.Session, url: str, path: str = None, query: dict = None) -> str:
-        """ Given the CDO endpoint, path, and query, return the json payload from the API """
+        """Given the CDO endpoint, path, and query, return the json payload from the API"""
+        # TODO: convert dictionary of query to encoded string with safe values..
+        # params = urllib.parse.quote(query.encode('utf-8'), safe='()/')
         uri = url if path is None else f"{url}/{path}"
-        result = http_session.get(url=uri, headers=http_session.headers, params=query)
+        result = http_session.get(
+            url=uri,
+            headers=http_session.headers,
+            params=query,
+        )
         result.raise_for_status()
         if result.text:
             return result.json()
@@ -71,7 +96,7 @@ class CDORequests:
     @CDOAPIWrapper()
     @staticmethod
     def post(http_session: requests.Session, url: str, path: str = None, data: dict = None, query: dict = None) -> str:
-        """ Given the CDO endpoint, path, and query, post the json data and return the json payload from the API """
+        """Given the CDO endpoint, path, and query, post the json data and return the json payload from the API"""
         uri = url if path is None else f"{url}/{path}"
         result = http_session.post(url=uri, params=query, json=data)
         result.raise_for_status()
@@ -83,7 +108,7 @@ class CDORequests:
     @CDOAPIWrapper()
     @staticmethod
     def put(http_session: requests.Session, url: str, path: str = None, data: dict = None, query: dict = None) -> str:
-        """ Given the CDO endpoint, path, and query, return the json payload from the API """
+        """Given the CDO endpoint, path, and query, return the json payload from the API"""
         uri = url if path is None else f"{url}/{path}"
         result = http_session.put(url=uri, headers=http_session.headers, params=query, json=data)
         result.raise_for_status()